+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Upgrade the python template dependencies
 - Upgrade the pyspark template to spark 3.3.1, also upgrade the dependencies
 
+### removals
+- Remove support for pyspark 2.x support from the pyspark template
+
 ## [1.1.1 - 2022-08-23]
 
 ### features